Split Util
